[F] Rewrite slider componentry for a11y
Replaces react-slick with a `useCarousel` hook that uses flickity under
the hood, and a series of components for rendering the carousel, prev &
next buttons, and pagination using accessible markup. Styles have been
modified slightly as well for better active/hover/focus appearance, but
we are still waiting on final designs for those states.

Resolves #50
Issue #55
Issue #56
Resolves #60
Resolves #63