[B] Fix rendering of dynamic list block items
Content blocks that let you show a list of other entries has “all” as
the default value for the number of entries to show. Due to a bug fixed
in 9acc406, this value was causing failed API requests for the entry
data, and so those entries were never rendered. After fixing that,
however, we are now requesting tons of entries on the frontend.
It's clear that it was a mistake to set "all" as the default value, and
due to these bugs content editors have been using "all" and assuming
that it doesn't actually show all entries. To fix this, in the client
we'll treat "all" as "none", and make a corresponding change in the
backend to rename that field for future block usage.